Redo Telemetry Augmentation
This changes the way how operators are augmented in order to produce
telemetry events. Instead of augmenting every RxJS operator individually
once they are imported, we patch Observable.pipe now. This allows us to
augment not only built-in, but also custom made operators as well.

Unfortunately, we cannot identify an operator solely by its source
location (line and character) anymore; The stack frame we generate
refers to the pipe function now. We can overcome this by adding the
argument index of the pipe function to the new IOperatorIdentifier.

This change reworks the TelemetryEvent type so it becomes extensible in
the future further. There is now a concrete OperatorLogPointTelemetry.
Additional telemetry events can easily be added later.